requirements.txt
Flask
pymongo
gunicorn
pytest
requests
werkzeug
markdown
flask-compress
stripe
openai
chromadb
google-generativeai
sentence-transformers
azure-cosmos
azure-storage-blob
azure-keyvault-secrets
azure-identity
azure-monitor-opentelemetry-exporter
